move file reading methods over to mixin, remove 'prefer uri' since path is the full uri
